+++ src/IntegratedDemonlist.hpp
@@ -16,8 +16,10 @@ public:
     inline static bool TRIED_LOADING = false;
 
     static void initializeDemons(web::WebResponse*, bool);
+    static void isOk(std::string const&, MiniFunction<void(bool, int)> callback);
+    static void isOk(std::string const&, EventListener<web::WebTask>&&, MiniFunction<void(bool, int)> callback);
     static void loadAREDL();
-    static void loadAREDL(EventListener<web::WebTask>&&, MiniFunction<void()> callback);
+    static void loadAREDL(EventListener<web::WebTask>&&, EventListener<web::WebTask>&&, LoadingCircle*, MiniFunction<void()> callback);
     static void loadPemonlist();
-    static void loadPemonlist(EventListener<web::WebTask>&&, MiniFunction<void()> callback);
+    static void loadPemonlist(EventListener<web::WebTask>&&, EventListener<web::WebTask>&&, LoadingCircle*, MiniFunction<void()> callback);
 };